+++ run-info/main.cpp
@@ -6,21 +6,17 @@
 
 using namespace geode::prelude;
 
-enum class Position {
-	Left,
-	Right
-};
-
-class RunInfoWidget : public CCNode {
+class RunInfoWidget : public CCNodeRGBA {
 public:
 	CCLabelBMFont* m_status_label = nullptr;
 	CCLabelBMFont* m_info_label = nullptr;
 	CCSprite* m_icon_sprite = nullptr;
+	CCNodeRGBA* m_top_layout = nullptr;
 	bool m_was_practice = true;
 
-	static RunInfoWidget* create(PlayLayer* layer, Position pos) {
+	static RunInfoWidget* create(PlayLayer* layer, bool left) {
 		auto* ret = new (std::nothrow) RunInfoWidget;
-		if (ret && ret->init(layer, pos)) {
+		if (ret && ret->init(layer, left)) {
 			ret->autorelease();
 			return ret;
 		} else {
@@ -29,17 +25,35 @@ public:
 		}
 	}
 
-	bool init(PlayLayer* layer, Position pos) {
-		if (!CCNode::init()) return false;
+	bool init(PlayLayer* layer, bool left) {
+		if (!CCNodeRGBA::init()) return false;
+		this->setCascadeColorEnabled(true);
+		this->setCascadeOpacityEnabled(true);
+
+		m_top_layout = CCNodeRGBA::create();
+		m_top_layout->setCascadeColorEnabled(true);
+		m_top_layout->setCascadeOpacityEnabled(true);
+		this->addChild(m_top_layout);
+		m_top_layout->setLayout(
+			AxisLayout::create(Axis::Row)
+			->setAutoScale(false)
+			->setGap(2.f)
+			->setAutoGrowAxis(true)
+			->setGrowCrossAxis(false)
+			->setCrossAxisOverflow(true)
+			->setAxisAlignment(AxisAlignment::Start)
+		);
+		m_top_layout->getLayout()->ignoreInvisibleChildren(true);
+		m_top_layout->setID("top-layout"_spr);
+
 		m_status_label = make_factory(CCLabelBMFont::create("Practice", "bigFont.fnt"))
-			.setOpacity(64)
 			.setScale(0.5f)
 			.setID("mode-label"_spr)
-			.addTo(this)
+			.addTo(m_top_layout)
+			.setZOrder(1)
 			.end();
 
 		m_info_label = make_factory(CCLabelBMFont::create("From 0%", "bigFont.fnt"))
-			.setOpacity(64)
 			.setScale(0.4f)
 			.setID("from-label"_spr)
 			.addTo(this)
@@ -48,11 +62,20 @@ public:
 		this->setID("RunInfoWidget"_spr);
 		this->reset_icon("checkpoint_01_001.png");
 
-		this->update_position(layer, pos);
+		this->setLayout(
+			AxisLayout::create(Axis::Column)
+			->setAutoScale(false)
+			->setGap(0.f)
+			->setAxisReverse(true)
+			->setAutoGrowAxis(true)
+			->setGrowCrossAxis(false)
+			->setCrossAxisOverflow(true)
+			->setCrossAxisAlignment(AxisAlignment::Start)
+			->setCrossAxisLineAlignment(AxisAlignment::Start)
+		);
+		this->getLayout()->ignoreInvisibleChildren(true);
 
-		float height = m_status_label->getScaledContentSize().height + m_info_label->getScaledContentSize().height;
-		float width = m_status_label->getScaledContentSize().width;
-		this->setContentSize({width, height});
+		this->update_position(layer, left);
 
 		return true;
 	}
@@ -60,16 +83,15 @@ public:
 	void reset_icon(const char* str) {
 		if (m_icon_sprite) m_icon_sprite->removeFromParent();
 		m_icon_sprite = make_factory(CCSprite::createWithSpriteFrameName(str))
-			.setOpacity(64)
 			.setID("icon-sprite"_spr)
-			.addTo(this)
+			.addTo(m_top_layout)
 			.with([&](auto* sprite) {
 				sprite->setScale(m_status_label->getScaledContentSize().height / sprite->getContentSize().height);
 			})
 			.end();
 	}
 
-	void update_position(PlayLayer* layer, Position pos) {
+	void update_position(PlayLayer* layer, bool left) {
 		bool show_icon = Mod::get()->getSettingValue<bool>("show-icon");
 		if (!Mod::get()->getSettingValue<bool>("use-start-pos") && !m_was_practice)
 			show_icon = false;
@@ -78,41 +100,33 @@ public:
 		// From x% is useless in platformer mode, force it off
 		const bool show_info = !is_platformer && Mod::get()->getSettingValue<bool>("show-from");
 
-		const float y = show_info ? m_info_label->getScaledContentSize().height : 0.f;
-		m_icon_sprite->setPositionY(y - 1.f);
-		m_status_label->setPositionY(y);
-		m_info_label->setPositionY(0.f);
-
-		const bool left = pos == Position::Left;
-		const float h_flip = left ? 1.f : -1.f;
-		const auto anchor = ccp(left ? 0.f : 1.f, 0.f);
-
-		m_icon_sprite->setAnchorPoint(anchor);
-		m_status_label->setAnchorPoint(anchor);
-		m_info_label->setAnchorPoint(anchor);
-
-		float x = 5.f;
-
-		m_icon_sprite->setPositionX(h_flip * x);
-		if (show_icon)
-			m_status_label->setPositionX(h_flip * (x + m_icon_sprite->getScaledContentSize().width + 2.f));
-		else
-			m_status_label->setPositionX(h_flip * x);
-		m_info_label->setPositionX(h_flip * x);
+		static_cast<AxisLayout*>(this->getLayout())->setCrossAxisLineAlignment(left ? AxisAlignment::Start : AxisAlignment::End);
+		static_cast<AxisLayout*>(m_top_layout->getLayout())->setAxisReverse(!left);
 
 		m_icon_sprite->setVisible(show_icon);
 		m_status_label->setVisible(show_status);
 		m_info_label->setVisible(show_info);
 
-		float height = 0.f;
-		if (show_info)
-			height += m_info_label->getScaledContentSize().height;
-		if (show_status || show_icon)
-			height += m_status_label->getScaledContentSize().height;
-		this->setContentSize({this->getContentSize().width, height});
+		m_top_layout->setVisible(m_icon_sprite->isVisible() || m_status_label->isVisible());
+
+		if (Mod::get()->getSettingValue<bool>("same-line") && !m_status_label->isVisible()) {
+			static_cast<AxisLayout*>(this->getLayout())
+				->setAxis(Axis::Row)
+				->setCrossAxisLineAlignment(AxisAlignment::Center)
+				->setAxisReverse(!left)
+				->setGap(2.f);
+		} else {
+			static_cast<AxisLayout*>(this->getLayout())
+				->setAxis(Axis::Column)
+				->setAxisReverse(true)
+				->setGap(0.f);
+		}
+
+		m_top_layout->updateLayout();
+		this->updateLayout();
 	}
 
-	void update_labels(PlayLayer* layer, int percent) {
+	void update_labels(PlayLayer* layer, float percent) {
 		if (layer->m_isPracticeMode) {
 			m_status_label->setString("Practice");
 			if (!m_was_practice)
@@ -125,7 +139,16 @@ public:
 			m_was_practice = false;
 		}
 
-		m_info_label->setString(fmt::format("From {}%", percent).c_str());
+		bool decimal = Mod::get()->getSettingValue<bool>("use-decimal");
+
+		if (decimal) {
+			m_info_label->setString(fmt::format("From {:.2f}%", percent).c_str());
+		} else {
+			m_info_label->setString(fmt::format("From {}%", int(percent)).c_str());
+		}
+
+		m_top_layout->updateLayout();
+		this->updateLayout();
 	}
 };
 
@@ -133,9 +156,7 @@ public:
 class $modify(PlayLayer) {
 	struct Fields {
 		RunInfoWidget* m_widget = nullptr;
-		// TODO: once we recreate getCurrentPercent maybe switch to float,
-		// and add option for decimals
-		int m_initial_percent = 0;
+		float m_initial_percent = 0;
 	};
 
 	bool init(GJGameLevel* level, bool unk1, bool unk2) {
@@ -157,12 +178,16 @@ class $modify(PlayLayer) {
 
 		const auto win_size = CCDirector::sharedDirector()->getWinSize();
 
-		m_fields->m_widget = make_factory(RunInfoWidget::create(this, Position::Left))
+		m_fields->m_widget = make_factory(RunInfoWidget::create(this, true))
 			.setAnchorPoint(ccp(0.f, 1.f))
 			.setPosition(0.f, win_size.height - 2.f)
 			.setZOrder(999)
 			.addTo(this)
 			.end();
+
+		m_fields->m_widget->setOpacity(Mod::get()->getSettingValue<int>("opacity"));
+		m_fields->m_widget->setScale(Mod::get()->getSettingValue<float>("scale"));
+		m_fields->m_widget->setColor(Mod::get()->getSettingValue<ccColor3B>("color"));
 
 		this->update_labels();
 		this->update_position();
@@ -183,28 +208,28 @@ class $modify(PlayLayer) {
 
 		auto* widget = m_fields->m_widget;
 
-		widget->update_position(this, left ? Position::Left : Position::Right);
+		widget->update_position(this, left);
 
 		const auto win_size = CCDirector::sharedDirector()->getWinSize();
 
 		if (left) {
-			widget->setPositionX(0.f);
+			widget->setPositionX(5.f);
 		} else {
-			widget->setPositionX(win_size.width);
+			widget->setPositionX(win_size.width - 5.f);
 		}
 
 		if (top) {
-			widget->setAnchorPoint(ccp(0.f, 1.f));
+			widget->setAnchorPoint(ccp(left ? 0.f : 1.f, 1.f));
 			widget->setPositionY(win_size.height - 2.f);
 		} else {
-			widget->setAnchorPoint(ccp(0.f, 0.f));
+			widget->setAnchorPoint(ccp(left ? 0.f : 1.f, 0.f));
 			widget->setPositionY(4.f);
 		}
 	}
 
 	void resetLevel() {
 		PlayLayer::resetLevel();
-		m_fields->m_initial_percent = this->getCurrentPercentInt();
+		m_fields->m_initial_percent = this->getCurrentPercent();
 		this->update_labels();
 		this->update_position();
 	}